fix a few int vs ssize_t bits around

+++ libknet/nsscrypto.h
@@ -9,14 +9,14 @@ struct crypto_instance;
 int crypto_authenticate_and_decrypt (
 	struct crypto_instance *instance,
 	unsigned char *buf,
-	int *buf_len);
+	ssize_t *buf_len);
 
 int crypto_encrypt_and_sign (
 	struct crypto_instance *instance,
 	const unsigned char *buf_in,
-	const size_t buf_in_len,
+	const ssize_t buf_in_len,
 	unsigned char *buf_out, 
-	size_t *buf_out_len);
+	ssize_t *buf_out_len);
 
 int crypto_init(
 	knet_handle_t knet_h,
